Break app/config out to config.go
